Conflicts:
	ios/sdk/WeexSDK.podspec
	ios/sdk/WeexSDK/Sources/Manager/WXComponentManager.m
	ios/sdk/WeexSDK/Sources/Module/WXPickerModule.h
	ios/sdk/WeexSDK/Sources/View/WXRootView.h

+++ ios/sdk/WeexSDK/Sources/Bridge/WXJSCoreBridge.m
@@ -149,7 +149,7 @@
         NSString *parentRef = [ref toString];
         NSInteger insertIndex = [[index toNumber] integerValue];
         
-         WXLogDebug(@"callAddElement...%@, %@, %@, %ld", instanceIdString, parentRef, componentData, insertIndex);
+         WXLogDebug(@"callAddElement...%@, %@, %@, %ld", instanceIdString, parentRef, componentData, (long)insertIndex);
         
         return callAddElement(instanceIdString, parentRef, componentData, insertIndex);
     };
